backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
mysql-connector-python>=8.0.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.0.0
passlib[bcrypt]>=1.7.0
python-decouple>=3.0
pydantic[email]>=2.0.0
pydantic-settings>=2.0.0
python-dateutil>=2.8.0
email-validator>=2.0.0
aiofiles>=0.8.0
pytest>=7.0.0
pytest-asyncio>=0.20.0
httpx>=0.24.0
coverage>=7.0.0
requests>=2.28.0 